requirements.txt
flask==2.3.2
requests==2.31.0
flask-cors==4.0.0
gunicorn==20.1.0
python-dotenv==1.0.0
numpy>=1.26.0
